Update Profile caching to use get_object_cached_field.
Djblets received a new function for fetching related objects from a
`select_related()` or `prefetch_related()` cache, which can simplify a
lot of code. One of the core tests for this was converting our somewhat
complex `User.get_profile()` cache logic to use that.

This change moves this method over to that, simplifying the function a
bit and helping ensure that functionality continues to work as expected.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14328/