Make it possible to archive repositories that are inaccessible.
The PUT /repositories/<n>/ API supports an `archive_name` field which
can hide the repository and give it a special name. This was going
through the regular repository form to actually handle the save
operation, which meant that if the repository being archived was
inaccessible for some reason, the operation would fail.

This change makes it so we check if the only thing being done is
archiving the repository, and if so we skip using the form and just
update the model itself.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14415/