Remove the newly-added allowDeclaredFields setting for Babel.
I recently added `allowDeclaredFields` so that I could utilize the
`declare` modifier for variables. However, with Babel, this isn't what
we want.

Normally, any variable defined in a class but not set to a value will be
simply left out of the prototype (an implicit `undefined`), but with
this setting enabled, only variables with `declare` will be left out
(everything else will be set to an explicit `undefined`).

That breaks checks like `this.prototype.hasOwnProperty(...)` that should
have been `false` if a variable was not explicitly set.

We can instead avoid `declare` and just define it, getting the same
results. This only really matters to us today for static variables
promoted to prototype attributes through Spina.

Testing Done:
Rebuilt all static media and verified the breakages in unit tests have
been fixed.

Reviewed at https://reviews.reviewboard.org/r/14406/