Add safer checks in ExtraDataMixin's logic to check extraData support.
`ExtraDataMixin` has a series of checks to determine if it should
set `this.extraData`. When calling `set()`, it makes a determination
based on the attributes passed in and the current object. This doesn't
always do the right thing, though.

If attributes are passed in and `this.extraData` is not set, then we
fail when trying to call a method on it.

Along with that, the method used to check `this.extraData` was a
presence check in the property, which isn't as safe as a truthiness
check. While fine in practice, this issue was encountered when changing
Babel's TypeScript options to enable `allowDeclaredFields`, which
resulted in `extraData` being explicitly set to `undefined`.

The safer check and the asserts should help catch any confusing issues
in the future.

Testing Done:
Unit tests pass.

Tested with the failure condition and saw the more useful error.

Reviewed at https://reviews.reviewboard.org/r/14407/