Add a setting for controlling whether to confirm a Ship It! review.
With the modern location for the Ship It action (in the Review menu),
it's much harder to accidentally post a Ship It review. This makes the
confirmation prompt less important than it once was.

We now have a setting for controlling whether to prompt for
confirmation. This is managed in My Account -> Settings.

The confirmation dialog itself now uses an Ink modal dialog, with some
additional text to help guide users. It also includes a "Do not ask
again" checkbox, which will turn off the confirmation setting as well.
This is a sort of prototype for a suppression dialog that we may want in
Ink.

The setting is managed through the User Session resource, which now
accepts a HTTP PUT with a `settings:json` JSON Merge Patch of settings.
`RB.UserSession` provides an endpoint for this as well. These currently
only accept this new `confirm_ship_it` setting, and they're marked as
for internal use only, but it provides a path for more dynamic settings
down the road.

Testing Done:
Python and JavaScript unit tests passed.

Tested the Ship It flow with confirmation turned on and off.

Tested the "Do not ask again" checkbox in the confirmation dialog,
verifying that it turned off the confirmation setting.

Reviewed at https://reviews.reviewboard.org/r/14409/