Fix positioning of the action menus when off the side of the screen.
Actions menus that are near the right-hand side of the screen end up
being positioned too far to the right. This is easily seen with the
Follow menu. The logic, copied from Ink, was incorrect in its
calculations, and ended up off by the difference between the window
width and the right edge of the handle element.

It now attempts to more carefully place this on the screen without
scrolling off the side. First, it tries to right-align it to the parent
handle. If that's still off the side, then it tries to align it to the
right-most pixel of the window. And then, if we're off the left-hand
side, it tries to align to the left-most pixel of the window.

This logic will also be copied over to Ink.

Testing Done:
Tested all the action menus I could find, and also tested with this
logic in Ink menus exhibiting the issue.

Reviewed at https://reviews.reviewboard.org/r/14419/