Use the local site aware session URL when updating User session settings.
We recently added new User session settings for customizing a Quick Access
hotbar for actions, and for toggling the Ship It\! confirmation dialog.
The session API endpoint now handles updating these settings. However,
when calling the User session API endpoint from our frontend, we failed
to plumb through the local site to the request URL. This lead to a 404
response on local site servers (such as on RBCommons).

This change passes the local site for the current review request through
to the API request. Now the User session settings can be customized as
intended on local site servers. For a user that has access to multiple
local sites, the settings are shared accross all the local sites.

Testing Done:
- Ran unit tests, but these currently fail due to issues unrelated to
  this change.

Tested the following on a local site, and a non-local site:
- Tested toggling the Ship It! confirmation.
- Tested customizing the Quick Access actions hotbar.

Reviewed at https://reviews.reviewboard.org/r/14427/