Fix unit tests for new ship-it dialog.
The unit test for the new ship-it dialog that attempted to change the
confirmation preference must have been written while the user sessions
settings code was still in flux, and wasn't properly skipping the code
that actually saved the setting. This resulted in a bad PUT to the
server and the ship-it dialog being permanently visible on the screen.

This change fixes the test so the `storeSettings()` method isn't
actually invoked, but resolves the promise, allowing things to proceed.

Testing Done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/14443/