Spruce up the integrations docs.
This change modernizes the screenshots and instructions in our integrations
docs. The instructions step for adding a new integration configuration
had outdated screenshots for each integration, which we replace with a
general screenshot showing the full list of integrations to choose from
when you click the "Add Integration" button.

We update the integrations conditions screenshot and add it to some
integration docs pages where it was missing. This is a good way to show
the different types of conditions that could be used for an integration.

We remove the screenshots that show just the “Name” and “Webhook URL” fields.
I find that they don’t add much value, they just show the empty text field
which is very easy to find without the screenshot.

Testing Done:
Built the docs and viewed them.

Reviewed at https://reviews.reviewboard.org/r/14448/