Add Doc Converter to the sample docker-compose.yaml configurations.
This updates our existing sample docker compose config files to show how to
launch Review Board alongside an instance of Doc Converter, the document
conversion service needed for Power Pack's document review feature.

The configs now have an instruction step for dealing with "optional services",
which are any services that we may include in the file but are optional. For
now RabbitMQ and Doc Converter are the only optional services, but in the
future we'll add Review Bot too.

The instructions for the RabbitMQ service make it clear that it can be
used with multiple services, not just Doc Converter, and lists the
services that require it.

Testing Done:
Tested this sample docker-compose and successfully ran the containers. Was
able to upload documents to the Review Board server and have them converted.

Reviewed at https://reviews.reviewboard.org/r/14199/