Fix bug infobox link URLs when using local sites.
The bug infobox was using `local_site_reverse` to create the link for
opening a bug, but was not passing the local site into that call. This
meant we'd get the wrong URL (or in the context of RBCommons, get an
error).

This change fixes that and adds some additional tests for the infobox,
including rendering with and without local sites.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14461/