Add a new cancelled state for status updates.
This change adds a new cancelled state that can be set for status
updates. At the moment, this displays the same as error states, but it
has a default description of "cancelled" (for updates that do not have
explicit description text set), and it does not show a "Retry" button
for services that allow manually triggering retries.

Testing Done:
- Ran unit tests.
- Used this in conjunction with some code in rbintegrations.

Reviewed at https://reviews.reviewboard.org/r/14466/