Improve rbext create boilerplate, add type hints.
The boilerplate created by `rbext create` was a little bit crufty,
having a mix of fake code and real. This change cleans that up so that
it's true boilerplate only, and all the fake stuff is in comments.

This change also modernizes the implementation of `rbext` with type
hints and some modern syntax nicety.

Testing Done:
Generated extensions with `rbext create`.

Reviewed at https://reviews.reviewboard.org/r/14543/